odrađeno dodavanje,brisanje i izmjena sponzora u potpunosti

+++ src/controllers/SponsorController.java
@@ -3,13 +3,9 @@ package controllers;
 import javafx.collections.FXCollections;
 import javafx.collections.transformation.FilteredList;
 import javafx.event.Event;
-import javafx.scene.input.KeyEvent;
 import javafx.scene.input.MouseEvent;
-import javafx.scene.layout.AnchorPane;
 import javafx.scene.layout.VBox;
 import model.DAO.DAOFactory;
-import model.DAO.mysql.MySQLSponsorDAO;
-import javafx.beans.Observable;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
@@ -29,38 +25,43 @@ public class SponsorController {
 
     @FXML
     private TextField sponsorNameField;
-
-
     @FXML
     private TableView<Sponsor> sponsorTable;
-
     private ObservableList<Sponsor> sponsorList;
-
-
-
     @FXML
     private SponsorSidebarController sponsorSidebarController;
     private AddEditSponsorController addEditSponsorController;
     private Stage addEditSponsorStage;
 
     @FXML
-    void addSponsor(ActionEvent event) throws Exception {
-        addEditSponsorStage.show();
+    void addSponsor(ActionEvent event) {
+        addEditSponsorController.setSponsor(new Sponsor());
+        addEditSponsorStage.showAndWait();
+        sponsorList=new FilteredList<Sponsor>(FXCollections.observableList(DAOFactory.getDAOFactory().getSponsorDAO().sponsors()));
+        sponsorTable.setItems(sponsorList);
     }
 
     @FXML
-    void changeSponsor(ActionEvent event) throws Exception {
-       addEditSponsorStage.show();
+    void changeSponsor(ActionEvent event)  {
+        if(sponsorTable.getSelectionModel().isEmpty())
+            return;
+        Sponsor selected=sponsorTable.getSelectionModel().getSelectedItem();
+        addEditSponsorController.setSponsor(selected);
+        addEditSponsorStage.showAndWait();
+        sponsorTable.refresh();
     }
 
     @FXML
-    void deleteSponsor(ActionEvent event) throws Exception {
+    void deleteSponsor(ActionEvent event) {
+        if(sponsorTable.getSelectionModel().isEmpty())
+            return;
         Sponsor selection=sponsorTable.getSelectionModel().getSelectedItem();
-        System.out.println(selection);
         if( selection!= null && DAOFactory.getDAOFactory().getSponsorDAO().delete(selection)){
-            sponsorList.remove(selection);
+            sponsorList=new FilteredList<Sponsor>(FXCollections.observableList(DAOFactory.getDAOFactory().getSponsorDAO().sponsors()));
+            sponsorTable.setItems(sponsorList);
         }
       else{
+            try{
             FXMLLoader loader=new FXMLLoader(this.getClass().getResource("../view/alert.fxml"));
             VBox alert=loader.load();
             AlertController alertController=loader.getController();
@@ -68,6 +69,7 @@ public class SponsorController {
             Stage alertStage=new Stage();
             alertStage.setScene(new Scene(alert));
             alertStage.show();
+            }catch (Exception e){e.printStackTrace();}
 
          }
 
@@ -77,9 +79,8 @@ public class SponsorController {
     void sponsorSearch(Event event) {
        ObservableList<Sponsor> filtered=sponsorList.filtered( e->e.getName().toLowerCase().matches(".*"+sponsorNameField.getText().toLowerCase()+".*"));
        sponsorTable.setItems(filtered);
-       if(filtered.size()==1){
-           sponsorSidebarController.setSponsor(filtered.get(0));
-       }
+       sponsorSidebarController.setSponsor(filtered.get(0));
+
     }
 
     @FXML
@@ -90,6 +91,7 @@ public class SponsorController {
     }
     public void setSponsorSidebarController(SponsorSidebarController sponsorSidebarController) {
         this.sponsorSidebarController = sponsorSidebarController;
+        sponsorSidebarController.setSponsor(sponsorList.get(0));
     }
 
     @FXML
@@ -105,7 +107,6 @@ public class SponsorController {
         sponsorTable.getColumns().get(0).setCellValueFactory(new PropertyValueFactory<>("name"));
         sponsorTable.getColumns().get(1).setCellValueFactory(new PropertyValueFactory<>("kind"));
         sponsorTable.getColumns().get(2).setCellValueFactory(new PropertyValueFactory<>("address"));
-
 
     }
 }
